split up plotting helpers
